Trabalho 01/requirements.txt
pysimplegui==4.7.0
pillow==8.1.0
numpy==1.20.1